requirements.txt
numpy
pandas
joblib
paho-mqtt
tensorflow
pyserial
google-auth